Start standalone library project
Originally this was written for a Logic World server. But once another project required it, I made this standalone version.

This project got refactored before the standalone port. A custom Exception system was replaced with Errors. Some recursion and reference counting got removed -- in general an improvement of code quality.